requirements.txt
# conda
pandas
ipython
# pip
openpyxl==3.0.7
nltk==3.6.2
tokenize-uk==0.2.0
requests==2.25.1
pystempel==1.2.0
git+https://github.com/Desklop/Uk_Stemmer
